requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-datatable==0.3.1
django-datatable-view==2.1.5
django-datatables-view==1.19.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Employee==0.0.4
Flask==1.1.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
legacy==0.1.6
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.1.2
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==1.0.1
whitenoise==5.3.0
